:sparkles: Move TimelineDataStore to `CAFU.Timeline.Data.DataStore`
